requirements.txt
slack-sdk>=3.9.1
everyaction-client>=0.3.0
